// src/redux/types/EventDetail.ts
export interface IEventDetail {
    sourceOfFunds: string;
    taxReference: string;
    natureOfIncome: string;
    placeOfDeath: string;
    nameOfHospital: string;
    drNursePathologist: string;
    drNursePathologistPracticeNumber: string;
    funeralParlour: string;
    chiefsName: string;
    chiefsPhoneNumber: string;
    dha1663: string;
    dha1680: string;
    deathCertificateIssuedByUser: string;
    haThatIssuedDeathCertificate: string;
  }